# internal/api/openapi.json
{
 "openapi": "3.0.3",
 "info": {
  "title": "Gorget REST API",
  "version": "1.0.0",
  "description": "The API behind the web console. Sign in with a session cookie (send X-CSRF-Token on writes) or use an API token: `Authorization: Bearer gat_...`. Tokens are read-only or read/write. IdP provisioning uses the separate SCIM 2.0 endpoint under /scim/v2."
 },
 "servers": [
  {
   "url": "/api/v1"
  }
 ],
 "security": [
  {
   "bearer": []
  },
  {
   "session": []
  }
 ],
 "components": {
  "securitySchemes": {
   "bearer": {
    "type": "http",
    "scheme": "bearer",
    "description": "API token created under Account > API tokens"
   },
   "session": {
    "type": "apiKey",
    "in": "cookie",
    "name": "gorget_session"
   }
  },
  "schemas": {
   "Error": {
    "type": "object",
    "properties": {
     "code": {
      "type": "string"
     },
     "message": {
      "type": "string"
     },
     "details": {}
    }
   },
   "Ok": {
    "type": "object",
    "properties": {
     "ok": {
      "type": "boolean"
     }
    }
   },
...
  }
 },
 "paths": {
  "/setup/status": {
   "get": {
    "tags": [
     "Setup"
    ],
    "summary": "Is first-run setup finished?",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "properties": {
          "completed": {
           "type": "boolean"
          },
          "version": {
           "type": "string"
          }
         }
        }
       }
      }
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "security": []
   }
  },
  "/setup": {
   "post": {
    "tags": [
     "Setup"
    ],
    "summary": "Finish first-run setup (needs the one-time setup token)",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      }
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "network_name": {
          "type": "string"
         },
         "owner_email": {
          "type": "string"
         },
         "owner_name": {
          "type": "string"
         },
         "password": {
          "type": "string"
         },
         "ipv4": {
          "type": "string"
         },
         "domain": {
          "type": "string"
         },
         "policy": {
          "type": "string"
         },
         "approval_required": {
          "type": "boolean"
         },
         "setup_token": {
          "type": "string"
         }
        },
        "required": [
         "owner_email",
         "password",
         "setup_token"
        ]
       }
      }
     }
    },
    "security": []
   }
  },
  "/auth/login": {
   "post": {
    "tags": [
     "Auth"
    ],
    "summary": "Sign in with email and password",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "properties": {
          "csrf_token": {
           "type": "string"
          },
          "mfa_required": {
           "type": "boolean"
          },
          "mfa_methods": {
           "type": "array",
           "items": {
            "type": "string"
           }
          }
         }
        }
       }
      }
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "email": {
          "type": "string"
         },
         "password": {
          "type": "string"
         }
        },
        "required": [
         "email",
         "password"
        ]
       }
      }
     }
    },
    "security": []
   }
  },
  "/auth/logout": {
   "post": {
    "tags": [
     "Auth"
    ],
    "summary": "Sign out",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      }
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   }
  },
  "/auth/providers": {
   "get": {
    "tags": [
     "Auth"
    ],
    "summary": "List single sign-on providers",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "properties": {
           "id": {
            "type": "string"
           },
           "name": {
            "type": "string"
           }
          }
         }
        }
       }
      }
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "security": []
   }
  },
  "/auth/oidc/{id}/start": {
   "get": {
    "tags": [
     "Auth"
    ],
    "summary": "Start an OIDC sign-in (redirects)",
    "responses": {
     "302": {
      "description": "OK"
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "security": []
   }
  },
  "/auth/mfa/totp": {
   "post": {
    "tags": [
     "Auth"
    ],
    "summary": "Complete sign-in with an authenticator or recovery code",
    "responses": {
     "200": {
      "description": "OK"
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "code": {
          "type": "string"
         }
        }
       }
      }
     }
    }
   }
  },
  "/me": {
   "get": {
    "tags": [
     "Account"
    ],
    "summary": "Current user, permissions and features",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/User"
        }
       }
      }
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "Account"
    ],
    "summary": "Update your profile",
    "responses": {
     "200": {
      "description": "OK"
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "name": {
          "type": "string"
         }
        }
       }
      }
     }
    }
   }
  },
  "/me/password": {
   "post": {
    "tags": [
     "Account"
    ],
    "summary": "Change your password",
    "responses": {
     "200": {
      "description": "OK"
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "current": {
          "type": "string"
         },
         "new": {
          "type": "string"
         }
        }
       }
      }
     }
    }
   }
  },
  "/me/sessions": {
   "get": {
    "tags": [
     "Account"
    ],
    "summary": "Your active sign-ins",
    "responses": {
     "200": {
      "description": "OK"
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   }
  },
  "/me/tokens": {
   "get": {
    "tags": [
     "Account"
    ],
    "summary": "Your API tokens",
    "responses": {
     "200": {
      "description": "OK"
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "Account"
    ],
    "summary": "Create an API token (shown once)",
    "responses": {
     "201": {
      "description": "OK"
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "name": {
          "type": "string"
         },
         "scopes": {
          "type": "array",
          "items": {
           "type": "string"
          }
         },
         "expires_in_days": {
          "type": "integer"
         }
        }
       }
      }
     }
    }
   }
  },
  "/me/tokens/{id}": {
   "delete": {
    "tags": [
     "Account"
    ],
    "summary": "Delete an API token",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      }
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   }
  },
  "/overview": {
   "get": {
    "tags": [
     "Dashboard"
    ],
    "summary": "Counts for the dashboard",
    "responses": {
     "200": {
      "description": "OK"
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   }
  },
  "/network-map": {
   "get": {
    "tags": [
     "Dashboard"
    ],
    "summary": "Devices and the connections the policy allows",
    "responses": {
     "200": {
      "description": "OK"
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   }
  },
  "/events": {
   "get": {
    "tags": [
     "Dashboard"
    ],
    "summary": "Server-sent events stream of changes",
    "responses": {
     "200": {
      "description": "OK"
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   }
  },
  "/devices": {
   "get": {
    "tags": [
     "Devices"
    ],
    "summary": "List devices",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Device"
         }
        }
       }
      }
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "schema": {
       "type": "string"
      },
      "required": false,
      "description": "search text"
     },
     {
      "name": "kind",
      "in": "query",
      "schema": {
       "type": "string"
      },
      "required": false,
      "description": "native|wireguard|gateway"
     }
    ]
   }
  },
  "/devices/{id}": {
   "get": {
    "tags": [
     "Devices"
    ],
    "summary": "One device",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Device"
        }
       }
      }
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   },
   "patch": {
    "tags": [
     "Devices"
    ],
    "summary": "Rename, tag, approve, disable, set key expiry or exit-node approval",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Device"
        }
       }
      }
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "name": {
          "type": "string"
         },
         "tags": {
          "type": "array",
          "items": {
           "type": "string"
          }
         },
         "ipv4": {
          "type": "string"
         },
         "key_expiry_disabled": {
          "type": "boolean"
         },
         "exit_approved": {
          "type": "boolean"
         },
         "state": {
          "type": "string"
         },
         "expires_at": {
          "type": "integer"
         },
         "tunnel_mode": {
          "type": "string"
         },
         "custom_allowed_ips": {
          "type": "array",
          "items": {
           "type": "string"
          }
         }
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   },
   "delete": {
    "tags": [
     "Devices"
    ],
    "summary": "Delete a device",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      }
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   }
  },
  "/devices/{id}/approve": {
   "post": {
    "tags": [
     "Devices"
    ],
    "summary": "Approve a pending device",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Device"
        }
       }
      }
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "description": "Needs: network admin",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   }
  },
  "/devices/{id}/expire-key": {
   "post": {
    "tags": [
     "Devices"
    ],
    "summary": "Force the device to sign in again",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      }
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   }
  },
  "/devices/{id}/access": {
   "get": {
    "tags": [
     "Devices"
    ],
    "summary": "What this device may reach, and who may reach it",
    "responses": {
     "200": {
      "description": "OK"
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   }
  },
  "/devices/{id}/wireguard-config": {
   "get": {
    "tags": [
     "Devices"
    ],
    "summary": "Config file for a standard WireGuard device",
    "responses": {
     "200": {
      "description": "OK"
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   }
  },
  "/devices/{id}/rotate-psk": {
   "post": {
    "tags": [
     "Devices"
    ],
    "summary": "Rotate the pre-shared key of a WireGuard device",
    "responses": {
     "200": {
      "description": "OK"
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   }
  },
  "/wireguard-configs": {
   "post": {
    "tags": [
     "Devices"
    ],
    "summary": "Create a standard WireGuard device (config and QR code)",
    "responses": {
     "201": {
      "description": "OK"
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "name": {
          "type": "string"
         },
         "user_id": {
          "type": "string"
         },
         "public_key": {
          "type": "string"
         },
         "tunnel_mode": {
          "type": "string"
         },
         "custom_allowed_ips": {
          "type": "array",
          "items": {
           "type": "string"
          }
         },
         "expires_at": {
          "type": "integer"
         },
         "preshared_key": {
          "type": "boolean"
         },
         "dns": {
          "type": "boolean"
         },
         "tags": {
          "type": "array",
          "items": {
           "type": "string"
          }
         },
         "ipv4": {
          "type": "string"
         }
        }
       }
      }
     }
    }
   }
  },
...
  "/geoip": {
   "get": {
    "tags": [
     "Settings"
    ],
    "summary": "Country database status",
    "responses": {
     "200": {
      "description": "OK"
     },
     "default": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   }
  },
...
 }
}